requirements-pi.txt
-r requirements.txt

# Voice hardware (requires Python 3.11: tflite-runtime has no aarch64 wheels for 3.12+)
sounddevice>=0.5
faster-whisper>=1.1
openwakeword>=0.6
piper-tts>=1.4
kokoro>=0.9.4
soundfile>=0.12
